// apps/portal/src/api/generated/model/setVoteRulesetDtoQuorumMeasure.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */

export type SetVoteRulesetDtoQuorumMeasure =
    (typeof SetVoteRulesetDtoQuorumMeasure)[keyof typeof SetVoteRulesetDtoQuorumMeasure];

export const SetVoteRulesetDtoQuorumMeasure = {
    UNIT_SHARE: "UNIT_SHARE",
    UNIT_COUNT: "UNIT_COUNT",
} as const;